feat: 数据质量页面开发完成 (Q-1 ~ Q-6)
第二十九轮数据质量页面功能开发：

Q-1 数据库表 + ORM 模型：
- ddl.sql 新增 data_quality_report 表
- models.py 新增 DataQualityReport ORM 类

Q-2 质量分析服务：
- services/data_quality.py 新建
- CompletenessChecker: 字段覆盖率检查
- FreshnessChecker: 数据新鲜度检查
- AnomalyDetector: 异常值检测
- QualityService: 评分聚合服务

Q-3 API 端点：
- GET /api/quality/report 获取最近报告
- POST /api/quality/check 触发质量检查
- GET /api/quality/history 获取历史记录

Q-4 8000 端口前端：
- templates/index.html 侧边栏新增导航
- 质量评分卡片（3类别 × 3维度）
- 异常详情展示区域

Q-5 Vue 前端同步：
- frontend/src/api/index.js 新增质量 API
- frontend/src/views/Quality.vue 新建页面
- frontend/src/router/index.js 新增路由
- frontend/src/App.vue 新增导航

Q-6 采集后自动触发：
- trigger_quality_check_after_collect() 函数
- 采集任务完成后自动质量检查
- WebSocket 推送质量更新通知

+++ services/data_validator.py
@@ -105,7 +105,7 @@ class DataValidator:
             return []
 
         check_fields = fields or list(df.columns)
-        total = len(df)
+        total = int(len(df))
         report = []
 
         for field_name in check_fields:
@@ -117,8 +117,8 @@ class DataValidator:
                     coverage_rate=0.0
                 ))
             else:
-                covered = df[field_name].notna().sum()
-                rate = covered / total if total > 0 else 0.0
+                covered = int(df[field_name].notna().sum())
+                rate = float(covered / total) if total > 0 else 0.0
                 report.append(FieldCoverage(
                     field_name=field_name,
                     total_count=total,
